Refact rpc system

+++ connection.go
@@ -8,7 +8,7 @@ type ConnectionService struct {
 	connCountLock sync.RWMutex // protect connCount
 	connCount     int
 	sessUUIDLock  sync.RWMutex // protect sessUUID
-	sessUUID      int
+	sessUUID      uint64
 }
 
 func NewConnectionService() *ConnectionService {
@@ -34,14 +34,14 @@ func (c *ConnectionService) getCurrentConnCount() int {
 	return c.connCount
 }
 
-func (c *ConnectionService) getNewSessionUUID() int {
+func (c *ConnectionService) getNewSessionUUID() uint64 {
 	c.sessUUIDLock.Lock()
 	defer c.sessUUIDLock.Unlock()
 	c.sessUUID++
 	return c.sessUUID
 }
 
-func (c *ConnectionService) getCurrentSessionUUID() int {
+func (c *ConnectionService) getCurrentSessionUUID() uint64 {
 	c.sessUUIDLock.RLock()
 	defer c.sessUUIDLock.RUnlock()
 	return c.sessUUID
